requirements.txt
boto3==1.34.0
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
google-auth-oauthlib==1.2.0
